pyproject.toml
[project]
name = "akvtaskws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.14",
    "aioredis>=2.0.1",
    "aiosmtplib>=4.0.0",
    "asyncio>=3.4.3",
    "logfire>=3.9.0",
    "pydantic>=2.10.6",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
    "websockets>=15.0.1",
]
